moved useRegistrationState.test into folder w/ the respective file
